# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "spelunk-server",
    "description": "Shared memory server for spelunk. Stores decisions, requirements, and context for a team and serves them over HTTP.",
    "contact": {
      "name": "spelunk",
      "url": "https://github.com/spelunk-cloud/spelunk"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/memory/harvested-shas": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Return all git commit SHAs stored in note tags for a project.",
        "description": "Each harvested commit is tagged `git:<sha>`. Clients call this endpoint\nto skip commits they have already stored, enabling incremental harvest.",
        "operationId": "harvested_shas",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of harvested git SHAs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/projects/{project_id}/memory/since": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Return notes newer than a cursor, in one of two modes:",
        "description": "- `?since_id=<uuid>`: delta-pull mode (wire parity with cloud-api;\n  `CloudSyncClient::pull_since`/`spelunk sync` targets this). Returns\n  `{entries, count}`, entries ordered by arrival at this server.\n- `?t=<unix_secs>`: legacy timestamp mode (`spelunk memory since`\n  targets this). Returns a bare array, ordered `created_at ASC`.\n\n`since_id` takes precedence when both are supplied. Archived entries are\nexcluded in both modes.",
        "operationId": "memory_since",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "t",
            "in": "query",
            "description": "Unix epoch seconds (exclusive lower bound). Required unless `since_id`\nis given; ignored when `since_id` is present.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "since_id",
            "in": "query",
            "description": "UUID cursor (exclusive lower bound, arrival-ordered; see\n`ServerDb::notes_since_id`). Takes precedence over `t` when both are\nsupplied. Selects the delta-pull response shape (`{entries, count}`)\ninstead of the bare array `t` returns.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default: 100, max: 500).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notes newer than `t` (bare array) or `since_id` (`{entries, count}`)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerNote"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Neither `t` nor `since_id` given",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/stream": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Stream new memory entries as Server-Sent Events. Each event carries a\nsingle `ServerNote` serialised as JSON. The stream polls the database once\nper second and stays open indefinitely; close the connection to stop it.",
        "description": "Pass `?t=<unix_secs>` to replay entries written after a known timestamp.\nOmit it to receive only entries written after the connection opens.",
        "operationId": "memory_stream",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "t",
            "in": "query",
            "description": "Unix epoch seconds to start from (inclusive). Defaults to now.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of new memory entries"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/projects/{project_id}/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Embed a search query server-side and return the vector for the CLI to use\nin its local KNN search.",
        "description": "The server applies the F2LLM code-retrieval query prefix\n(`Instruct: Given a code search query…\\nQuery: {q}`) so the CLI does not\nneed to know the embedding format. The server does **not** perform the KNN\nstep: the local SQLite index lives on the CLI side.\n\n- `\"semantic\"` / `\"hybrid\"`: embeds the query and returns `query_vector`.\n  Returns **400** if no embedder is configured on this server.\n- `\"text\"`: returns `query_vector: null`; the CLI falls back to FTS.",
        "operationId": "project_search",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Query vector (CLI runs KNN locally)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "No embedder configured or invalid mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Embed admission queue full; retry after the given delay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeSearchRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/projects/{project_id}/search`.",
        "required": [
          "query"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "description": "Maximum number of results the caller intends to fetch.\nPassed back in the response for informational purposes only: the server\ndoes not perform the KNN step; the CLI does that against its local index.",
            "minimum": 0
          },
          "mode": {
            "type": "string",
            "description": "Search mode: `\"hybrid\"` (default), `\"semantic\"`, or `\"text\"`.\n\n`\"hybrid\"` and `\"semantic\"` require an embedder; the server will return\n`query_vector` so the CLI can run KNN against its local SQLite index.\n`\"text\"` skips embedding and signals the CLI to use FTS."
          },
          "query": {
            "type": "string",
            "description": "Natural-language search query."
          }
        }
      },
      "CodeSearchResponse": {
        "type": "object",
        "description": "Response body for `POST /v1/projects/{project_id}/search`.",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "description": "Mode actually used (`\"semantic\"`, `\"hybrid\"`, or `\"text\"`).\nMay differ from the requested mode if the embedder is unavailable."
          },
          "query_vector": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Query embedding vector: present for semantic/hybrid modes.\nThe CLI uses this to run KNN against its local index.\n`null` when mode is `\"text\"` (no embedding needed)."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Consistent JSON error body: `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ServerNote": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "tags",
          "linked_files",
          "created_at",
          "status"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation."
          },
          "distance": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Cosine distance from query (only present in search results)."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "type": "string",
            "description": "Kind: `decision`, `requirement`, `note`, `question`, `handoff`, or `intent`."
          },
          "linked_files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "remote_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Canonical cross-machine id (uuid). Optional and additive; `None` for\nrows never assigned one. Absent on the wire when `None`."
          },
          "status": {
            "type": "string",
            "description": "`active` or `archived`."
          },
          "superseded_by": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Pass as `Authorization: Bearer <key>`. Not required when no key is configured on the server."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}